ODE-651 updated changes to allow proper use of Builder classes.

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetB10Builder.java
@@ -8,7 +8,7 @@ public class OffsetB10Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetB10(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(2).intValue();
+   public static Long offsetB10(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(2).longValue();
    }
 }
