+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetLLB16Builder.java
@@ -8,7 +8,7 @@ public class OffsetLLB16Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetLLB16(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(7).intValue();
+   public static Long offsetLLB16(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(7).longValue();
    }
 }
