+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetLLB18Builder.java
@@ -8,7 +8,7 @@ public class OffsetLLB18Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetLLB18(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(7).intValue();
+   public static Long offsetLLB18(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(7).longValue();
    }
 }
